google-adk-agents/jira-agent/requirements.txt
google-adk
jira
